+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.37.0
+
+* `[strutil]` Added helper `Q` for working with default values
+* `[usage]` Added color customization for application name and version
+
 ### 12.36.0
 
 * `[system]` Added CPU architecture bits info to `SystemInfo`